app/requirements.txt
Flask==2.0.1
tensorflow==2.6.0
pillow==8.3.1
gunicorn==20.1.0